+++ src/base/VolumeInterface.php
@@ -209,6 +209,18 @@ interface VolumeInterface extends SavableComponentInterface
      * @throws VolumeObjectNotFoundException if a directory with such name already exists
      * @throws VolumeObjectExistsException if a directory with such name already exists
      * @throws VolumeException if something else goes wrong
+     * @deprecated in 3.6.0. Use [[renameDirectory()]] instead.
      */
     public function renameDir(string $path, string $newName);
+
+    /**
+     * Renames a directory.
+     *
+     * @param string $path The path of the directory, relative to the source’s root
+     * @param string $newName The new path of the directory, relative to the source’s root
+     * @throws VolumeObjectNotFoundException if a directory with such name already exists
+     * @throws VolumeObjectExistsException if a directory with such name already exists
+     * @throws VolumeException if something else goes wrong
+     */
+    public function renameDirectory(string $path, string $newName);
 }