+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added `craft\base\VolumeInterface::deleteDirectory()`.
 - Added `craft\base\VolumeInterface::getDateModified()`.
 - Added `craft\base\VolumeInterface::getFileSize()`.
+- Added `craft\base\VolumeInterface::renameDirectory()`.
 - Added `craft\fieldlayoutelements\BaseField::useFieldset()`.
 - Added `craft\fields\Url::TYPE_EMAIL`.
 - Added `craft\fields\Url::TYPE_TEL`.
@@ -40,6 +41,7 @@
 - Deprecated `craft\base\VolumeInterface::createDir()`. `createDirectory()` should be used instead.
 - Deprecated `craft\base\VolumeInterface::deleteDir()`. `deleteDirectory()` should be used instead.
 - Deprecated `craft\base\VolumeInterface::getFileMetadata()`. `getFileSize()` and `getDateModified()` should be used instead.
+- Deprecated `craft\base\VolumeInterface::renameDir()`. `renameDirectory()` should be used instead.
 - Deprecated `craft\helpers\Template::paginateCriteria()`. `paginateQuery()` should be used instead.
 
 ### Removed